server/requirements.txt
alembic==1.11.1
bcrypt==4.0.1
bidict==0.22.1
blinker==1.6.2
certifi==2023.7.22
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.0.4
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
minio==7.1.16
packaging==23.2
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-dotenv==1.0.0
python-engineio==4.7.1
python-socketio==5.9.0
pytz==2023.3
schedule==1.2.1
simple-websocket==1.0.0
SQLAlchemy==2.0.15
typing_extensions==4.6.3
urllib3==2.0.4
websocket-client==1.5.1
Werkzeug==2.3.4
wsproto==1.2.0
WTForms==3.1.1
